% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGGArrows.R
\name{addGGArrows}
\alias{addGGArrows}
\title{Add Gradient Vector Arrows to Gradient Field Heatmap.}
\usage{
addGGArrows(
  g,
  centers,
  gradients,
  fac = 0.025,
  arrow_len = unit(0.075/2, "inches"),
  nColumns = 10L,
  nRows = 10L,
  ...
)
}
\arguments{
\item{g}{\link{\code{ggplot}}\cr
\code{ggplot}-object as created by \code{\link{ggplotHeatmap}}.}

\item{centers}{\link{\code{\link{matrix}}}\cr
Matrix containing the points of a grid of cells, for which the lengths of cumulated
gradient paths should be computed. Each row corresponds to a single point.}

\item{gradients}{\link{\code{\link{matrix}}}\cr
Matrix containing the multi-objective gradients for each point of \code{centers}.}

\item{fac}{\link{\code{\link{numeric}(1L)}}\cr
Factor used for scaling the lengths of the gradient vectors. The default is \code{0.025}.}

\item{arrow_len}{\link{\code{\link{unit}(1L)}}\cr
Length of the arrow tips (default: \code{unit(0.075 / 2, "inches")}).}

\item{nColumns}{\link{\code{\link{integer}(1L)}}\cr
How many columns of arrows should be drawn? The default is \code{10}.}

\item{nRows}{\link{\code{\link{integer}(1L)}}\cr
How many rows of arrows should be drawn? The default is \code{10}.}

\item{...}{\link{any}\cr
Further arguments to be passed to the \code{geom_tile} function of \code{ggplot}.}
}
\value{
\link{\code{ggplot}}\cr
A \code{ggplot} object displaying the multi-objective gradient landscape.
}
\description{
Add vectors indicating the gradient field to the heatmap of the
cumulated path lengths.
}
\examples{
# Define a bi-objective test problems and a grid of points:
fn = function(x) c(sum((x - c(0.2, 1))^2), sum((x - c(0.5, 0.5))^2))
points = as.matrix(expand.grid(x1 = seq(0, 0.7, 0.005), x2 = seq(0, 1.25, 0.005)))

# Compute the corresponding gradients and the cumulated path lengths:
gradients = computeGradientField(points, fn)
x = computeCumulatedPathLengths(points, gradients)

# Visualize the resulting multi-objective "landscape":
g = ggplotHeatmap(x)
g

# Add white arrows of the gradient field to the plot:
addGGArrows(g, points, gradients, color = "white")
}
