% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGradientField.R
\name{computeGradientField}
\alias{computeGradientField}
\title{Compute the multi-objective gradient vector for a set of points.}
\usage{
computeGradientField(
  points,
  fn,
  prec.grad = 1e-06,
  prec.norm = 1e-06,
  prec.angle = 1e-04,
  parallelize = FALSE,
  impute.boundary = TRUE,
  lower = NULL,
  upper = NULL,
  verbose = TRUE
)
}
\arguments{
\item{points}{[\link{matrix}]\cr
Matrix of points, for which the multi-objective gradient should be computed.
Each row of the matrix will be considered as a separate point, thus the number
of rows corresponds to the number of observations and the number of columns
to the dimensionality of the search space.}

\item{fn}{[\code{\link{function}}]\cr
Multi-objective function with up to three objectives.}

\item{prec.grad}{[\code{\link{numeric}}]\cr
Precision value (= step size) used for approximating the gradient. The default is \code{1e-6}.}

\item{prec.norm}{[\code{\link{numeric}}]\cr
Precision threshold when normalizing a vector. That is, every element of the vector,
whose absolute value is below this threshold, will be replaced by 0.
The default is \code{1e-6}.}

\item{prec.angle}{[\code{\link{numeric}}]\cr
Precision threshold used for comparing whether the angle (in degree) between two
vectors is zero. The default is \code{1e-4}.}

\item{parallelize}{[\code{\link{logical}}]\cr
Should the computation of the gradient vectors be parallelized (with \code{parallel::mclapply})?
The default is \code{FALSE}.}

\item{impute.boundary}{[\code{\link{logical}}]\cr
Should values at the boundary be imputed? This is relevant for problems
that have locally efficient points on the boundary box.}

\item{lower}{[\code{\link{numeric}}]\cr
Vector of lower bounds.}

\item{upper}{[\code{\link{numeric}}]\cr
Vector of upper bounds.}
}
\value{
[\code{\link{matrix}}]\cr
Returns \code{matrix} of multi-objective gradients. The i-th row of the matrix
contains the multi-objective gradient vector of the i-th observation (= row)
of the input matrix \code{points}.
}
\description{
Computes the multi-objective gradients for a matrix of \code{points}.
}
