% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConjPlot.R
\name{ConjPlot}
\alias{ConjPlot}
\title{Create combined plots for interaction contrasts}
\usage{
ConjPlot(results, modellino, sig.level = 0.05, colonne = 2, focus = FALSE)
}
\arguments{
\item{results}{A list of contrast data frame}

\item{modellino}{A fitted model object}

\item{sig.level}{Significance level threshold. Default is 0.05}

\item{colonne}{Number of columns in the patchwork layout. Default is 2}

\item{focus}{Logical: if \code{TRUE}, prints each plot separately; if \code{FALSE}, returns a combined plot.}
}
\description{
This function is for internal usage
}
\keyword{internal}
