% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomDescription.R
\name{CustomDescription}
\alias{CustomDescription}
\title{Print Descriptive Summaries for Custom Contrasts (Internal Use)}
\usage{
CustomDescription(output_contrasti, sig.level = 0.05, digits = 2)
}
\arguments{
\item{output_contrasti}{A data frame output}

\item{sig.level}{Significance level}

\item{digits}{Number of decimal places}
}
\value{
No return value.
}
\description{
Intended for internal usage
}
\keyword{internal}
