% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractMatrices.R
\name{ExtractMatrices}
\alias{ExtractMatrices}
\title{Extract Contrast and Hypothesis Matrices from a Model}
\usage{
ExtractMatrices(model, data)
}
\arguments{
\item{model}{A fitted model object}

\item{data}{A data frame used to fit the model}
}
\value{
A named list
}
\description{
For internal usage.
}
\keyword{internal}
