% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractParams.R
\name{ExtractParams}
\alias{ExtractParams}
\title{Extract Effect Sizes and Standardized Coefficients from a Regression Model}
\usage{
ExtractParams(model)
}
\arguments{
\item{model}{A fitted regression model object.}
}
\value{
A data frame.
}
\description{
Calculates Cohen's d and standardized coefficients.
}
\keyword{internal}
