% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOrthogonal.R
\name{FindOrthogonal}
\alias{FindOrthogonal}
\title{Identify Factor Variables with Orthogonal Contrasts}
\usage{
FindOrthogonal(model)
}
\arguments{
\item{model}{A fitted model object}
}
\value{
A character vector of variable names whose contrast matrices are orthogonal.
}
\description{
For internal usage
}
\keyword{internal}
