% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPCP.R
\name{ggplotGlobalPCP}
\alias{ggplotGlobalPCP}
\title{Global Multi-Objective PCP Visualizations}
\usage{
ggplotGlobalPCP(
  obj.space,
  scale = "uniminmax",
  alphaLines = 0.2,
  check.data = TRUE
)
}
\arguments{
\item{obj.space}{[\code{\link{matrix}}]\cr}

\item{scale}{[\code{\link{character}}]\cr}

\item{alphaLines}{[\code{\link{numeric}}]\cr}

\item{check.data}{[\code{\link{logical}}] \cr
Should sanity checks be performed? The default is \code{TRUE}. Note that the checks
should only be turned off (e.g., for a slight speed up), if you are sure that you
provide the input data in the correct format.}
}
\description{
Returns a \strong{Global PCP} (\strong{P}arallel \strong{C}oordinate \strong{P}lot) visualization
as a \link[ggplot2:ggplot]{ggplot2::ggplot} object for the given visualization data. Global PCPs
only show the Pareto front, i.e., the globally efficient points.
}
