% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotHeatmap.R
\name{ggplotHeatmap}
\alias{ggplotHeatmap}
\title{Visualize Heatmap of Multi-Objective Gradients.}
\usage{
ggplotHeatmap(
  df,
  var1 = "x1",
  var2 = "x2",
  log.scale = TRUE,
  impute.zero = TRUE,
  minimalistic.image = FALSE,
  color.palette,
  legend.position,
  ...
)
}
\arguments{
\item{df}{\link{\code{\link{data.frame}}}\cr
Data frame as returned by \code{\link{computeCumulatedPathLengths}}.}

\item{var1}{\link{\code{\link{character}(1L)}}\cr
Name of the variable indicating the first dimension (default: \code{"x1"}).}

\item{var2}{\link{\code{\link{character}(1L)}}\cr
Name of the variable indicating the second dimension (default: \code{"x2"}).}

\item{log.scale}{\link{\code{\link{logical}(1L)}}\cr
Should the resulting heights be displayed on a log-scale? The default is \code{TRUE}.}

\item{impute.zero}{\link{\code{\link{logical}(1L)}}\cr
Should height values, which are exactly zero be imputed by a value half the magnitude
of the smallest non-zero height? Otherwise ggplot will automatically color the
corresponding tiles by a color representing \code{NA} values (usually grey).
Note that this parameter is only relevant in case of \code{log.scale = TRUE}.
The default is \code{TRUE}.}

\item{minimalistic.image}{\link{\code{\link{logical}(1L)}}\cr
Should all information surrounding the image (axes, legends, background, etc.) be discarded?
The default is \code{FALSE}.}

\item{color.palette}{\link{\code{\link{character}}}\cr
Vector of colors used for visualizing the different heights of the landscape. By default,
this function tries to use the color palettes from \code{fields::tim.color} or
\code{viridis}. However, if neither of these packages is installed, it will use
\code{terrain.colors}.}

\item{legend.position}{\link{\code{\link{character}(1L)}}\cr
On which side of the plot, should the legend be located? If this information is not provided
and \code{minimalisitic.image = FALSE}, the legend will be placed on the right side.}

\item{...}{\link{any}\cr
Further arguments to be passed to the \code{geom_tile} function of \code{ggplot}.}
}
\value{
\link{\code{ggplot}}\cr
A \code{ggplot} object displaying the multi-objective gradient landscape.
}
\description{
Visualization of the multi-objective gradient landscape by means of
a heatmap and on the basis of the \code{ggplot2}-package.
}
\examples{
# Define two single-objective test problems and a grid of points:
fn1 = function(x) sum((x - c(0.2, 1))^2)
fn2 = function(x) sum((x - c(0.5, 0.5))^2)
points = as.matrix(expand.grid(x1 = seq(0, 0.7, 0.005), x2 = seq(0, 1.25, 0.005)))

# Compute the corresponding gradients and the cumulated path lengths:
gradients = computeGradientField(points, fn1, fn2)
x = computeCumulatedPathLengths(points, gradients)

# Finally, we can visualize the resulting multi-objective "landscape":
ggplotHeatmap(x)

# If one is only interested in the image itself, but not in any labels, legends, etc.
# one should set minimalistic.image = TRUE:
ggplotHeatmap(x, minimalistic.image = TRUE)
}
