% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPCP.R
\name{ggplotLocalPCP}
\alias{ggplotLocalPCP}
\title{Local Multi-Objective PCP Visualizations}
\usage{
ggplotLocalPCP(
  design,
  less,
  space = c("objective", "decision", "both"),
  scale = "uniminmax",
  alphaLines = 0.1,
  check.data = TRUE
)
}
\arguments{
\item{design}{[\code{\link{list}}]\cr}

\item{less}{[\code{\link{list}}]\cr}

\item{space}{\code{"objective" || "decision" || "both"}}

\item{scale}{[\code{\link{character}}]\cr}

\item{alphaLines}{[\code{\link{numeric}}]\cr}

\item{check.data}{[\code{\link{logical}}] \cr
Should sanity checks be performed? The default is \code{TRUE}. Note that the checks
should only be turned off (e.g., for a slight speed up), if you are sure that you
provide the input data in the correct format.}
}
\description{
Returns a \strong{Local PCP} (\strong{P}arallel \strong{C}oordinate \strong{P}lot) visualization
as a \link[ggplot2:ggplot]{ggplot2::ggplot} object for the given visualization data. Local PCPs
highlight the different locally efficient sets.
}
