% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPLOT.R
\name{ggplotPLOT}
\alias{ggplotPLOT}
\title{Create a PLOT Visualization}
\usage{
ggplotPLOT(dec.space, obj.space, sinks, height, check.data = TRUE)
}
\arguments{
\item{dec.space}{Numeric \link[base:matrix]{base::matrix} that defines the evaluated points in decision space.
One column per dimension.}

\item{obj.space}{Numeric \link[base:matrix]{base::matrix} that defines the evaluated points in objective space.
One column per objective.}

\item{sinks}{Integer \link[base:vector]{base::vector} of (row) indices that identify the \emph{locally efficient} points.}

\item{height}{Numeric \link[base:vector]{base::vector} that assigns a height value to each evaluated point.}

\item{check.data}{\link[base:logical]{base::logical}: Whether to check data types of the other arguments.}
}
\description{
Returns a PLOT visualization as a \link[ggplot2:ggplot]{ggplot2::ggplot} object for the given visualization data.
}
