% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotSetTransitions.R
\name{ggplotSetTransitions}
\alias{ggplotSetTransitions}
\title{Set Transition Visualization in ggplot2}
\usage{
ggplotSetTransitions(design, less)
}
\arguments{
\item{design}{[\code{\link{list}}] \cr}

\item{less}{[\code{\link{list}}] \cr}
}
\value{
A [\link[ggplot2:ggplot]{ggplot2::ggplot}] object
}
\description{
Set Transition Visualization in ggplot2
}
\examples{

}
