% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupContrastCol.R
\name{GroupContrastCol}
\alias{GroupContrastCol}
\title{Group Contrast Columns by Interaction Structure}
\usage{
GroupContrastCol(contrast_matrix, data, model)
}
\arguments{
\item{contrast_matrix}{A design or contrast matrix}

\item{data}{The original data frame used to fit the model}

\item{model}{A fitted model object}
}
\value{
A named list
}
\description{
For internal usage
}
\keyword{internal}
