% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpretationScore.R
\name{InterpretationScore}
\alias{InterpretationScore}
\title{Compute Interpretation Score for Categorical Contrasts}
\usage{
InterpretationScore(variabile, tol = 1e-08)
}
\arguments{
\item{variabile}{A factor or categorical variable}

\item{tol}{A numeric tolerance (default \code{1e-8})}
}
\value{
An integer interpretation score
}
\description{
For internal usage
}
\keyword{internal}
