% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAsparFunction.R
\name{makeAsparFunction}
\alias{makeAsparFunction}
\title{Multi-Objective Aspar Function}
\usage{
makeAsparFunction(dimensions = 2, n.objectives = 2)
}
\arguments{
\item{dimensions}{[\code{\link{numeric}}]\cr
Number of dimensions}

\item{n.objectives}{[\code{\link{numeric}}]\cr
Number of objectives}
}
\value{
Multi-objective smoof function
}
\description{
Multi-Objective Aspar Function
}
