% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBiRosenbrockFunction.R
\name{makeBiRosenbrockFunction}
\alias{makeBiRosenbrockFunction}
\title{Bi-Objective Rosenbrock Function}
\usage{
makeBiRosenbrockFunction()
}
\value{
Multi-objective smoof function
}
\description{
Creates one particular parametrization of a bi-objective Rosenbrock function.
}
\examples{

}
