% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExtendedBiObjBBOBFunction.R
\name{makeExtendedBiObjBBOBFunction}
\alias{makeExtendedBiObjBBOBFunction}
\title{Smoof generator function for the extended bi-objective BBOB
with validated (corrected) instance mapping.}
\usage{
makeExtendedBiObjBBOBFunction(dimensions, fid, iid)
}
\description{
Smoof generator function for the extended bi-objective BBOB
with validated (corrected) instance mapping.
}
