% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMinDistFunction.R
\name{makeMinDistFunction}
\alias{makeMinDistFunction}
\alias{makeBiObjMinDistFunction}
\title{Multi-Objective MinDist Functions}
\usage{
makeMinDistFunction(
  centers.f1 = list(c(-2, -1), c(2, 1)),
  centers.f2 = list(c(-2, 1), c(2, -1)),
  centers.f3 = list(c(0, 1), c(0, -1))
)

makeBiObjMinDistFunction(
  centers.f1 = list(c(-2, -1), c(2, 1)),
  centers.f2 = list(c(-2, 1), c(2, -1))
)
}
\arguments{
\item{centers.f1}{[\code{\link{list}}]\cr List of centers for the first objective.}

\item{centers.f2}{[\code{\link{list}}]\cr List of centers for the second objective.}

\item{centers.f3}{[\code{\link{list}}] (optional)\cr List of centers for the third
objective, if function has three objectives.}
}
\value{
A multi-objective smoof function
}
\description{
Creates a multi-objective MinDist problem. Centers a defined for each
function with the \code{centers.fi} arguments. Box constraints are created at with
a distance of 1 to the extremal values of the centers.
}
\details{
Decision space dimensionality is determined by the length of each center
value in the \code{centers.fi} lists. If \code{is.null(centers.f3)}, the problem is
bi-objective, otherwise it is tri-objective. For creating bi-objective
problems, you can also use \code{makeBiObjMinDistFunction()} directly.
}
