% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRudolph9to9.R
\name{makeRudolph9to9Function}
\alias{makeRudolph9to9Function}
\title{Rudolph 9-to-9}
\usage{
makeRudolph9to9Function()
}
\value{
Multi-objective smoof function
}
\description{
Creates the Rudolph 9-to-9 test function
}
\examples{

}
