% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSGKFunction.R
\name{makeSGKFunction}
\alias{makeSGKFunction}
\title{Multi-Objective SGK Function}
\usage{
makeSGKFunction(dimensions = 2)
}
\arguments{
\item{dimensions}{\code{c(2, 3)}\cr
Number of (decision space) dimensions.}
}
\value{
Multi-objective smoof function
}
\description{
Multi-Objective SGK Function
}
