% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelContrastTable.R
\name{ModelContrastTable}
\alias{ModelContrastTable}
\title{Visualizing the Structure of a Linear Model}
\usage{
ModelContrastTable(LinearModel, type = "all")
}
\arguments{
\item{LinearModel}{A linear model object created using \code{lm()}.}

\item{type}{A character string specifying the type of table to return. Options include:
\itemize{
\item \code{"all"} (default): Returns both model-level and variable-level summaries.
\item \code{"summary"}: Returns the model-level summary.
\item \code{"x&y"}: Returns the variable-level summary (including dependent and independent variables).
\item \code{"x"}: Returns a summary of independent variables only.
\item \code{"y"}: Returns a summary of the dependent variable only.
}}
}
\value{
A table or a list of tables summarizing the structure of the model, depending on the \code{type} argument:
\itemize{
\item For \code{"all"}: A list containing both model-level and variable-level summary tables.
\item For \code{"summary"}: A data frame summarizing overall model structure.
\item For \code{"x&y"}: A data frame with variable names, types, levels, and contrast types.
\item For \code{"x"} or \code{"y"}: A data frame focused on the specified subset of variables.
}
}
\description{
This function takes a linear model object created with the \code{lm()} function and returns tabular summaries describing the model structure. It reports the number and types of variables, levels of factors, and the type of contrast matrices applied to categorical predictors.
}
\examples{
# Creating a linear model with the generative function
DT <- GenerateDataset1IV(n_sample = 30, means = c(120, 135, 115), sds = 20)
DT$iv <- as.factor(DT$iv)
model <- lm(dv ~ iv, data = DT)

# Without specifing the additional parameters (All tables by default)
ModelContrastTable(model)

# Function with only indipendent variables informations
ModelContrastTable(model, type = "x")

}
