% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotContrast.R
\name{PlotContrast}
\alias{PlotContrast}
\title{Internal plotting function for interaction contrasts}
\usage{
PlotContrast(contrast_data, pvalue, sig)
}
\arguments{
\item{contrast_data}{A data frame}

\item{pvalue}{A numeric value indicating the p-value of the interaction effect.}

\item{sig}{Significance level threshold (e.g., 0.05).}
}
\value{
A ggplot object representing the interaction contrast plot.
}
\description{
This function is for internal usage
}
