% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDashboard.R
\name{runDashboard}
\alias{runDashboard}
\title{Run the moPLOT dashboard}
\usage{
runDashboard()
}
\description{
\code{runDashboard} starts a \href{https://CRAN.R-project.org/package=shiny}{\code{shiny}}
application, which allows the interactive exploration of the optimization landscape
of many integrated benchmark suites.
}
