% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGGArrows.R
\name{addGGArrows}
\alias{addGGArrows}
\title{Add Gradient Vector Arrows to Gradient Field Heatmap.}
\usage{
addGGArrows(
  g,
  centers,
  gradients,
  fac = 0.025,
  arrow_len = unit(0.075/2, "inches"),
  nColumns = 10L,
  nRows = 10L,
  ...
)
}
\arguments{
\item{g}{[\code{\link[ggplot2:ggplot]{ggplot2::ggplot}}]\cr
\code{ggplot}-object as created by \code{\link{ggplotHeatmap}}.}

\item{centers}{[\code{\link{matrix}}]\cr
Matrix containing the points of a grid of cells, for which the lengths of cumulated
gradient paths should be computed. Each row corresponds to a single point.}

\item{gradients}{[\code{\link{matrix}}]\cr
Matrix containing the multi-objective gradients for each point of \code{centers}.}

\item{fac}{[\verb{[numeric]}]\cr
Factor used for scaling the lengths of the gradient vectors. The default is \code{0.025}.}

\item{arrow_len}{[\verb{[unit]}]\cr
Length of the arrow tips (default: \code{unit(0.075 / 2, "inches")}).}

\item{nColumns}{[\verb{[integer]}]\cr
How many columns of arrows should be drawn? The default is \code{10}.}

\item{nRows}{[\verb{[integer]}]\cr
How many rows of arrows should be drawn? The default is \code{10}.}

\item{...}{\link{any}\cr
Further arguments to be passed to the \code{geom_tile} function of \code{ggplot}.}
}
\value{
\code{\link{ggplot}}\cr
A \code{ggplot} object displaying the multi-objective gradient landscape.
}
\description{
Add vectors indicating the gradient field to the heatmap of the
cumulated path lengths.
}
