% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotPLOT.R
\name{ggplotPLOT}
\alias{ggplotPLOT}
\title{Create a PLOT Visualization}
\usage{
ggplotPLOT(
  dec.space,
  obj.space,
  sinks,
  height,
  colorscale.efficient = NULL,
  check.data = TRUE
)
}
\arguments{
\item{dec.space}{Numeric \link[base:matrix]{base::matrix} that defines the evaluated points in decision space.
One column per dimension.}

\item{obj.space}{Numeric \link[base:matrix]{base::matrix} that defines the evaluated points in objective space.
One column per objective.}

\item{sinks}{Integer \link[base:vector]{base::vector} of (row) indices that identify the \emph{locally efficient} points.}

\item{height}{Numeric \link[base:vector]{base::vector} that assigns a height value to each evaluated point.}

\item{check.data}{[\code{\link{logical}}] \cr
Should sanity checks be performed? The default is \code{TRUE}. Note that the checks
should only be turned off (e.g., for a slight speed up), if you are sure that you
provide the input data in the correct format.}
}
\description{
Returns a PLOT visualization as a \link[ggplot2:ggplot]{ggplot2::ggplot} object for the given visualization data.
}
