% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMPM2Function.R
\name{makeBiObjMPM2Function}
\alias{makeBiObjMPM2Function}
\alias{makeTriObjMPM2Function}
\title{Multi-Objective MPM2 Functions}
\usage{
makeBiObjMPM2Function(
  dimensions = 2,
  n.peaks.1 = 3,
  topology.1 = "random",
  seed.1 = 4,
  n.peaks.2 = 3,
  topology.2 = "random",
  seed.2 = 8
)

makeTriObjMPM2Function(
  dimensions = 2,
  n.peaks.1 = 3,
  topology.1 = "random",
  seed.1 = 4,
  n.peaks.2 = 3,
  topology.2 = "random",
  seed.2 = 8,
  n.peaks.3 = 3,
  topology.3 = "random",
  seed.3 = 12
)
}
\arguments{
\item{dimensions}{[\code{\link{numeric}}]\cr
Number of (decision space) dimensions of the MPM2 problems.}

\item{n.peaks.1}{[\code{\link{numeric}}]\cr
Number of peaks for the first constituent MPM2 problem.}

\item{topology.1}{\code{"random" || "funnel"}\cr
Topology of the first constituent MPM2 problem.}

\item{seed.1}{[\code{\link{numeric}}]\cr
Random seed of the first constituent MPM2 problem.}

\item{n.peaks.2}{[\code{\link{numeric}}]\cr
Number of peaks for the second constituent MPM2 problem.}

\item{topology.2}{\code{"random" || "funnel"}\cr
Topology of the second constituent MPM2 problem.}

\item{seed.2}{[\code{\link{numeric}}]\cr
Random seed of the second constituent MPM2 problem.}

\item{n.peaks.3}{[\code{\link{numeric}}]\cr
Number of peaks for the third constituent MPM2 problem.}

\item{topology.3}{\code{"random" || "funnel"}\cr
Topology of the third constituent MPM2 problem.}

\item{seed.3}{[\code{\link{numeric}}]\cr
Random seed of the third constituent MPM2 problem.}
}
\value{
Multi-objective smoof function
}
\description{
The \code{makeBiObjMPM2Function()} and \code{makeTriObjMPM2Function()} functions allow
the creation of multi-objective functions based on the MPM2 generator
implemented in \code{\link[smoof:makeMPM2Function]{smoof::makeMPM2Function()}}.
}
\examples{

}
